@@ -3,63 +3,25 @@
   "version": "1.0.0",
   "description": "A Vue.js project",
   "author": "Volkmar Kantor <volkmar.kantor@gmx.de>",
-  "license" : "GPL-3.0-or-later",
+  "license": "GPL-3.0-or-later",
+  "type": "module",
   "scripts": {
-    "dev": "cross-env NODE_ENV=development node build/dev-server.js",
+    "dev": "vite",
+    "build": "vite build",
+    "watch": "vite build --watch",
     "start": "node server.js",
-    "build": "node build/build.js",
     "run-json": "json-server --watch local.db.json --port 80"
   },
   "dependencies": {
-    "bootstrap": "^3.3.7",
-    "jquery": "^3.3.1",
-    "json-server": "^0.11.2",
-    "vue": "^2.5.16",
-    "vue-router": "^3.0.1"
+    "bootstrap": "^5.3.3",
+    "json-server": "^0.17.4",
+    "vue": "^3.4.30",
+    "vue-router": "^4.4.0"
   },
   "devDependencies": {
-    "autoprefixer": "^6.7.2",
-    "babel-core": "^6.26.3",
-    "babel-loader": "^6.2.10",
-    "babel-plugin-transform-runtime": "^6.22.0",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-stage-2": "^6.22.0",
-    "babel-register": "^6.26.0",
-    "chalk": "^1.1.3",
-    "connect-history-api-fallback": "^1.5.0",
-    "copy-webpack-plugin": "^4.5.1",
-    "cross-env": "^5.2.0",
-    "css-loader": "^0.28.11",
-    "eventsource-polyfill": "^0.9.6",
-    "express": "^4.16.3",
-    "extract-text-webpack-plugin": "^2.1.2",
-    "file-loader": "^0.11.1",
-    "friendly-errors-webpack-plugin": "^1.7.0",
-    "html-webpack-plugin": "^2.30.1",
-    "http-proxy-middleware": "^0.17.3",
-    "opn-cli": "^3.1.0",
-    "optimize-css-assets-webpack-plugin": "^1.3.0",
-    "ora": "^1.4.0",
-    "rimraf": "^2.6.2",
-    "semver": "^5.5.0",
-    "shelljs": "^0.7.6",
-    "url-loader": "^0.5.8",
-    "vue-loader": "^12.2.2",
-    "vue-style-loader": "^3.1.2",
-    "vue-template-compiler": "^2.5.16",
-    "webpack": "^2.7.0",
-    "webpack-bundle-analyzer": "^2.13.1",
-    "webpack-dev-middleware": "^1.12.2",
-    "webpack-hot-middleware": "^2.22.2",
-    "webpack-merge": "^4.1.3"
-  },
-  "engines": {
-    "node": ">= 4.0.0",
-    "npm": ">= 3.0.0"
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not ie <= 8"
-  ]
+    "@vitejs/plugin-vue": "^5.0.5",
+    "sass": "^1.77.6",
+    "sass-loader": "^14.2.1",
+    "vite": "^5.3.1"
+  }
 }
