@@ -1,15 +1,8 @@
-// The Vue build version to load with the `import` command
-// (runtime-only or standalone) has been set in webpack.base.conf with an alias.
-import Vue from 'vue';
+import { createApp } from 'vue'
 import router from './router';
-import App from './App';
-import Style from './assets/css/style.css';
+import './assets/css/style.css';
+import App from './App.vue'
 
-Vue.config.productionTip = false
-new Vue({
-  el: '#app',
-  router,
-  //components:{App},
-  //template: '<App></App>'
-  render: h => h(App)
-})
+createApp(App)
+  .use(router)
+  .mount('#app')
