@@ -5,12 +5,12 @@
         <h2>Bitte abholen</h2>
         <h1 :values="currentValues.length">
           <transition-group name="numberAnimated">
-            <span v-for="i in currentValues" :key="i">{{i.id}}</span>
+            <span v-for="i in currentValues" :key="i">{{i.number}}</span>
           </transition-group>
         </h1>
       </div>
     </transition>
-    
+
     <div class="topMenu">
       <a class="btn btn-primary btn-lg" @click="toggleFullscreen()"><span class="glyphicon glyphicon-resize-full" aria-hidden="true"></span></a>
       <router-link :to="{ name: 'home' }" class="btn btn-primary btn-lg">&times;</router-link>
@@ -19,7 +19,6 @@
 </template>
 
 <script>
-import $ from "jquery";
 import targetServer from '../services/target-server';
 import fullscreen from '../services/fullscreen';
 
@@ -27,63 +26,58 @@ export default {
   data () {
     return {
       currentValues: [],
-      continueRefresh: true,
-      refreshTimeout: null
+      intervalObject: null,
+      intervalDuration: 500,
     }
   },
-  mounted: function () {
-    this.refresh();
+  mounted() {
+    this.doSetup();
   },
-  beforeDestroy: function() {
+  beforeDestroy() {
     console.log('clearTimeout monitor');
     if(this.refreshTimeout) clearTimeout(this.refreshTimeout);
   },
   computed: {},
   methods: {
-    toggleFullscreen:function() {
+    doSetup() {
+      this.intervalObject = setInterval(() => {
+          this.loadData();
+        },this.intervalDuration);
+      this.loadData();
+    },
+    toggleFullscreen() {
       if(fullscreen.isFullscreen()) {
         fullscreen.exitFullscreen();
         return;
       }
       fullscreen.doFullscreen();
     },
-    doClose:function() {
+    doClose() {
       this.exitFullscreen();
       this.$emit('close');
     },
-    refresh: function() {
-      var that = this;
-      if(this.continueRefresh) {
-        $.ajax({
-          type:'GET',
-          url: targetServer.get()+'currentEntries',
-          success:function(response) {
-            that.handleResponse(response);
-          },
-          complete: function(jqXHR) {
-            //console.info('refresh done');
-            that.refreshTimeout = setTimeout(function() {
-              that.refresh();
-            },1000);
-          }
-        })
-      } else {
-        if(that.refreshTimeout) clearTimeout(that.refreshTimeout);
-      }
+    loadData() {
+      fetch(targetServer.get()+'currentEntries', {
+        method:'GET'
+      })
+      .then(response => response.json())
+      .then(data => {
+          this.handleResponse(data);
+        });
     },
-    handleResponse: function(response) {
+    handleResponse(response) {
       //remove not existing
-      for (var i = 0; i < this.currentValues.length; i++) {
+      for(var i = 0; i < this.currentValues.length; i++) {
         var current = this.currentValues[i];
-        if(response.find(function(value) { return value.id===current.id })===undefined) {
-          this.currentValues.splice(i,1);
+        if(response.find(value => value.id===current.id) === undefined) {
+          this.currentValues.splice(i, 1);
         }
       }
 
       //add new
       for (var i = 0; i < response.length; i++) {
         var current = response[i];
-        if(this.currentValues.find(function(value) { return value.id===current.id })===undefined) {
+        if(this.currentValues.find(value => value.id===current.id)===undefined) {
           this.currentValues.push(current);
         }
       }
