@@ -1,21 +1,26 @@
+import Home from '@/pages/Home.vue';
+import Monitor from '@/pages/Monitor.vue';
+import Controllroom from '@/pages/Controllroom.vue';
+import Page404 from '@/pages/404.vue';
+
+
 export default [
-    { 
-        path: '/', 
-        name:'home',
-        component: require('~/pages/Home.vue'),
-    }, {
-        path: '/monitor',
-        name: 'monitor',
-        component: require('~/pages/Monitor.vue'),
-    }, {
-        path: '/controll',
-        name: 'controll',
-        component: require('~/pages/Controllroom.vue'),
-    },
-    
-    { 
-        path: '*', 
-        component: require('~/pages/404.vue'), 
-        meta: { bodyClass: 'page404' }, 
-    }
+  {
+    path: '/',
+    name:'home',
+    component: Home,
+  }, {
+    path: '/monitor',
+    name: 'monitor',
+    component: Monitor,
+  }, {
+    path: '/controll',
+    name: 'controll',
+    component: Controllroom,
+  },
+  {
+    path: '/:pathMatch(.*)',
+    component: Page404,
+    meta: { bodyClass: 'page404' },
+  }
 ];
